requirements.txt
streamlit
google-generativeai
langchain
langchain_google_genai
haystack-ai
